remove console.log, improve placement
