+++ src/components/Layout/Layout.module.css
@@ -24,9 +24,8 @@ body {
 
 .topRight {
   display: flex;
-  gap: 0.5rem;
+  gap: 1rem;
   align-items: center;
-  flex-direction: row-reverse;
 }
 
 .topRight button {
