+++ src/contexts/auth.ts
@@ -21,8 +21,6 @@ const doAuth = async (callback: AuthCallbackFn, options: AuthOptions) => {
 
   // Get auth session cookie from preview subdomain
   chrome.cookies.get({ url: previewUrl, name: AUTH_SESSION_COOKIE_NAME }, function (cookie) {
-    console.log("cookie", cookie);
-
     if (cookie) {
       chrome.cookies.set({
         url: apiUrl!,
